refactor : moved coarsen function and added unit tests
