+++ ocf_data_sampler/load/utils.py
@@ -42,27 +42,3 @@ def get_xr_data_array_from_xr_dataset(ds: xr.Dataset) -> xr.DataArray:
     datavars = list(ds.var())
     assert len(datavars) == 1, "Cannot open as xr.DataArray: dataset contains multiple variables"
     return ds[datavars[0]]
-
-
-def coarsen_data(xr_data: xr.Dataset, coarsen_to_deg: float=0.1):
-    """
-    Coarsen the data to a specified resolution in degrees.
-    
-    Args:
-        xr_data: xarray dataset to coarsen
-        coarsen_to_deg: resolution to coarsen to in degrees
-    """
-
-    if "latitude" in xr_data.coords and "longitude" in xr_data.coords:
-        step = np.abs(xr_data.latitude.values[1]-xr_data.latitude.values[0])
-        step = np.round(step,4)
-        coarsen_factor = int(coarsen_to_deg/step)
-        if coarsen_factor > 1:
-            xr_data = xr_data.coarsen(
-                latitude=coarsen_factor,
-                longitude=coarsen_factor,
-                boundary="pad",
-                coord_func="min"
-            ).mean()
-        
-    return xr_data
